Refactor Middle component to use destructuring in props.tsx
